control_plane_api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-cloud-storage
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-logging
google-cloud-pubsub
firebase-admin
pandas
statsmodels
scikit-learn
pydantic
sqlalchemy
psycopg2-binary
cloud-sql-python-connector
httpx
pandas-gbq
numpy
scipy